requirements.txt
asgiref==3.4.1
Django==3.2.14
django-autoslug==1.9.8
django-ckeditor==6.3.2
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.0.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==5.3.0